chore: Move icon components out of public/ into components/
public/ is for static assets served at URL paths, not source code.
FactorIcons.tsx and InfoCircle.tsx were React components importable
via @/public/... — move them to components/ alongside InfoButton.tsx
and rewrite the six import sites (three app components, three tests).
