+++ app/[locale]/[state]/analytics/components/output-window.tsx
@@ -8,8 +8,8 @@ import {
   GovtResponse,
   RiskScore,
   Vulnerability,
-} from '@/public/FactorIcons';
-import { InfoSquare } from '@/public/InfoCircle';
+} from '@/components/FactorIcons';
+import { InfoSquare } from '@/components/InfoCircle';
 import { useQuery } from '@tanstack/react-query';
 import { useQueryState } from 'next-usequerystate';
 import { Button, Icon, Text, Tooltip } from 'opub-ui';
