+++ components/FactorIcons.tsx
